+++ MarkBitmap/MarkBitmap.cs
@@ -1,9 +1,5 @@
 ﻿using System;
-using System.Diagnostics;
 using System.Drawing;
-using System.Drawing.Imaging;
-using System.Runtime.InteropServices;
-using System.Runtime.Versioning;
 
 namespace MarkBitmap
 {
@@ -63,23 +59,94 @@ namespace MarkBitmap
             #region Vertical lines
 
             //
-            buffer = MarkVerticalLine(buffer, bitmap.Width / 3, 0, bitmap.Width, bitmap.Height - 1, color);
-            buffer = MarkVerticalLine(buffer, bitmap.Width * 2 / 3, 0, bitmap.Width, bitmap.Height - 1, color);
+            buffer = MarkVerticalLine(buffer, bitmap.Width / 3, 0, bitmap.Width, bitmap.Height, color);
+            buffer = MarkVerticalLine(buffer, bitmap.Width * 2 / 3, 0, bitmap.Width, bitmap.Height , color);
 
             #endregion Vertical lines
 
             #region Horizontal lines
 
             //
-            buffer = MarkHorizontalLine(buffer, 0, bitmap.Height / 3, bitmap.Width, bitmap.Width - 1, color);
-            buffer = MarkHorizontalLine(buffer, 0, bitmap.Height * 2 / 3, bitmap.Width, bitmap.Width - 1, color);
+            buffer = MarkHorizontalLine(buffer, 0, bitmap.Height / 3, bitmap.Width, bitmap.Width, color);
+            buffer = MarkHorizontalLine(buffer, 0, bitmap.Height * 2 / 3, bitmap.Width, bitmap.Width, color);
 
             #endregion Horizontal lines
 
             #region Corners
 
             //
-            buffer = MarkCorners(buffer, bitmap.Width, bitmap.Height, (bitmap.Width > bitmap.Height ? bitmap.Height / 10 : bitmap.Width / 10), color);
+            buffer = MarkCorners(buffer, bitmap.Width, bitmap.Height, bitmap.Width > bitmap.Height ? bitmap.Height / 12 : bitmap.Width / 12, color);
+
+            #endregion Corners
+
+            #region Center grid
+
+            //
+            int lengthX = bitmap.Width / 6;
+            int lengthY = bitmap.Height / 6;
+
+            //
+            int x = (bitmap.Width / 2) - (lengthX / 2);
+            int y = (bitmap.Height / 2) - (lengthY / 2);
+
+            //
+            int xOffset = x + lengthX;
+            int yOffset = y + lengthY;
+
+            //
+            buffer = MarkVerticalLine(buffer, x, y, bitmap.Width, lengthY, color);
+            buffer = MarkVerticalLine(buffer, xOffset, y, bitmap.Width, lengthY, color);
+
+            //
+            buffer = MarkHorizontalLine(buffer, x, y, bitmap.Width, lengthX, color);
+            buffer = MarkHorizontalLine(buffer, x, yOffset, bitmap.Width, lengthX, color);
+
+            #endregion Center grid
+
+            #region Center cross
+
+            //
+            int crossLength = bitmap.Width > bitmap.Height ? bitmap.Height / 12 : bitmap.Width / 12;
+
+            //
+            int xCross = bitmap.Width / 2;
+            int yCross = bitmap.Height / 2;
+
+            //
+            buffer = MarkVerticalLine(buffer, xCross - 1, yCross - (crossLength / 2), bitmap.Width, crossLength, color);
+            buffer = MarkVerticalLine(buffer, xCross, yCross - (crossLength / 2), bitmap.Width, crossLength, color);
+            buffer = MarkVerticalLine(buffer, xCross + 1, yCross - (crossLength / 2), bitmap.Width, crossLength, color);
+
+            //
+            buffer = MarkHorizontalLine(buffer, xCross - (crossLength / 2), yCross - 1, bitmap.Width, crossLength, color);
+            buffer = MarkHorizontalLine(buffer, xCross - (crossLength / 2), yCross, bitmap.Width, crossLength, color);
+            buffer = MarkHorizontalLine(buffer, xCross - (crossLength / 2), yCross + 1, bitmap.Width, crossLength, color);
+
+            #endregion Center cross
+
+            // Create marked bitmap.
+            Bitmap markedBitmap = ToBMP(buffer, bitmap.Width, bitmap.Height);
+
+            // Returning bitmap.
+            return markedBitmap;
+        }
+
+        public static Bitmap MarkTargetGrid(Bitmap bitmap, Color color)
+        {
+            // Checking if bitmap is null.
+            if (bitmap == null)
+            {
+                // Throwing an ArgumentNullException with specified message.
+                throw new ArgumentNullException($"nullass {bitmap}");
+            }
+
+            // Transform to byte array.
+            byte[] buffer = ToBuffer(bitmap);
+
+            #region Corners
+
+            //
+            buffer = MarkCorners(buffer, bitmap.Width, bitmap.Height, bitmap.Width > bitmap.Height ? bitmap.Height / 12 : bitmap.Width / 12, color);
 
             #endregion Corners
 
@@ -89,8 +156,8 @@ namespace MarkBitmap
             int length = bitmap.Width > bitmap.Height ? bitmap.Height / 6 : bitmap.Width / 6;
 
             //
-            int x = (bitmap.Width / 2) - length / 2;
-            int y = (bitmap.Height / 2) - length / 2;
+            int x = (bitmap.Width / 2) - (length / 2);
+            int y = (bitmap.Height / 2) - (length / 2);
 
             //
             int xOffset = x + length;
@@ -106,26 +173,12 @@ namespace MarkBitmap
 
             #endregion Center grid
 
-            #region Center cross
+            #region Diags
 
-            //
-            int crossLength = bitmap.Width > bitmap.Height ? bitmap.Height / 12 : bitmap.Width / 12;
+            buffer = MarkDiagonalLine(buffer, x, y, bitmap.Width, length, color);
+            buffer = MarkDiagonalInverseLine(buffer, xOffset, y, bitmap.Width, length, color);
 
-            //
-            int xCross = (bitmap.Width / 2);
-            int yCross = (bitmap.Height / 2);
-
-            //
-            buffer = MarkVerticalLine(buffer, xCross - 1, yCross - crossLength / 2, bitmap.Width, crossLength, Color.Yellow);
-            buffer = MarkVerticalLine(buffer, xCross, yCross - crossLength / 2, bitmap.Width, crossLength, color);
-            buffer = MarkVerticalLine(buffer, xCross + 1, yCross - crossLength / 2, bitmap.Width, crossLength, Color.Yellow);
-
-            //
-            buffer = MarkHorizontalLine(buffer, xCross - crossLength / 2, yCross - 1, bitmap.Width, crossLength, Color.Yellow);
-            buffer = MarkHorizontalLine(buffer, xCross - crossLength / 2, yCross, bitmap.Width, crossLength, color);
-            buffer = MarkHorizontalLine(buffer, xCross - crossLength / 2, yCross + 1, bitmap.Width, crossLength, Color.Yellow);
-
-            #endregion Center cross
+            #endregion Diags
 
             // Create marked bitmap.
             Bitmap markedBitmap = ToBMP(buffer, bitmap.Width, bitmap.Height);